+++ src/swap/central/swapuz.ts
@@ -22,13 +22,13 @@ import {
 
 import { div18 } from '../../util/biggystringplus'
 import {
-  checkInvalidCodes,
+  checkInvalidTokenIds,
   checkWhitelistedMainnetCodes,
   CurrencyPluginIdSwapChainCodeMap,
   ensureInFuture,
   getCodesWithTranscription,
   getMaxSwappable,
-  InvalidCurrencyCodes,
+  InvalidTokenIds,
   isLikeKind,
   makeSwapPluginQuote,
   SwapOrder
@@ -58,16 +58,22 @@ const asInitOptions = asObject({
 const orderUri = 'https://swapuz.com/order/'
 const uri = 'https://api.swapuz.com/api/home/v1/'
 
-const INVALID_CURRENCY_CODES: InvalidCurrencyCodes = {
+const INVALID_TOKEN_IDS: InvalidTokenIds = {
   from: {
-    ethereum: ['MATH'],
-    optimism: ['VELO'],
-    polygon: ['USDC', 'USDC.e']
+    ethereum: ['08d967bb0134f2d07f7cfb6e246680c53927dd30' /* MATH */],
+    optimism: ['9560e827af36c94d2ac33a39bce1fe78631088db' /* VELO */],
+    polygon: [
+      '3c499c542cef5e3811e1192ce70d8cc03d5c3359' /* USDC */,
+      '2791bca1f2de4661ed88a30c99a7a9449aa84174' /* USDC.e */
+    ]
   },
   to: {
-    ethereum: ['MATH'],
-    polygon: ['USDC', 'USDC.e'],
-    zcash: ['ZEC'],
+    ethereum: ['08d967bb0134f2d07f7cfb6e246680c53927dd30' /* MATH */],
+    polygon: [
+      '3c499c542cef5e3811e1192ce70d8cc03d5c3359' /* USDC */,
+      '2791bca1f2de4661ed88a30c99a7a9449aa84174' /* USDC.e */
+    ],
+    zcash: [null],
     zksync: 'allCodes'
   }
 }
@@ -156,7 +162,7 @@ export function makeSwapuzPlugin(opts: EdgeCorePluginOptions): EdgeSwapPlugin {
   ): Promise<SwapOrder> => {
     const { fromWallet, toWallet } = request
 
-    checkInvalidCodes(INVALID_CURRENCY_CODES, request, swapInfo)
+    checkInvalidTokenIds(INVALID_TOKEN_IDS, request, swapInfo)
     checkWhitelistedMainnetCodes(MAINNET_CODE_TRANSCRIPTION, request, swapInfo)
 
     // Grab addresses: