+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - changed: Replace wallet denomination methods with synchronous utils
+- changed: Use tokenIds in invalid currency objects
 
 ## 2.40.0 (2025-12-04)